[refactor] 🐳 빌드 lint 문제 해결 완료
