+++ life_on_hana/app/(routes)/home/lumpsum/page.tsx
@@ -13,6 +13,7 @@ import {
   type TLikedLoanProductDetailItemProps,
   type TRecommendItemProps,
 } from '@/types/componentTypes';
+import { Reason } from '@/types/dataTypes';
 import { RecommendItem } from '@/components/molecules/RecommendItem';
 import {
   fetchAccountSalary,
@@ -22,19 +23,7 @@ import {
 } from '@/api';
 import LikedLoanProductDetailItem from '@/components/molecules/LikedLoanProductDetailItem';
 import { useToast } from '@/hooks/use-toast';
-import WhilickItemLoading from '@/components/molecules/WhilickItemLoading';
 import LoadingIcon from '@/components/atoms/LoadingIcon';
-
-export enum Reason {
-  CHILDREN = '자녀 지원 (결혼, 학비, 독립 지원 등)',
-  MEDICAL = '의료비 지원 (본인 및 가족 의료비 등)',
-  HOUSING = '주거 및 생활비 (주거, 생활비 부족 등)',
-  BUSINESS_INVESTMENT = '사업 및 투자 자금 (투자, 창업 자금 등)',
-  VEHICLE_TRANSPORT = '차량 및 교통',
-  LEISURE = '여가 (여행, 취미, 교육 등)',
-  DEBT_REPAYMENT = '채무 상환',
-  OTHER = '기타',
-}
 
 type TSelectedProductProps = {
   type: 'LOAN';
